Fix travis build
Travis was failing because of some linting issue.

I have to create a common API file to share these methods across which
is also a good practice in case we've change one of the method in
sequential file it wouldn't cause bug in the other.
